# perf: add index to optimize jwt query