--- a/persistence/sql/migrations/20250520000001000000_add_oauth2_trusted_jwt_bearer_issuer_index.down.sql
+++ b/persistence/sql/migrations/20250520000001000000_add_oauth2_trusted_jwt_bearer_issuer_index.down.sql
@@ -1,0 +1,1 @@
+DROP INDEX hydra_oauth2_trusted_jwt_bearer_issuer_issuer_subject_key_id_key_2;