@alnr 2025-05-20
I'll review this tomorrow, esp. multiregion perf.